Split frequencies and behavior in Stochastic Language
Most of the time, we require the frequencies and variants separately,
not as a list of (variant, frequency) tuples.

So, split these into a StochasticLanguage object that has a list of
variants and an array of the frequencies (both have the same length).

The relative frequency of the n-th item in the variants list is the n-th
item in the frequencies array.