Add variant arg to import_log, add rustxes and update pm4py
pm4py update to support the rustxes import variant.

Need to keep pm4py <=2.7.10.1 because in 2.7.11.1, they update
`sample_cases`, which makes it random without the ability to set a seed.
You still could make the sampling deterministic by setting a property on
the dataframe, but I don't want to do that.